+++ src/main/java/com/github/larseckart/tcr/ArlosGitNotationPrompt.java
@@ -1,31 +1,18 @@
 package com.github.larseckart.tcr;
 
-import java.awt.Container;
-import java.awt.Dimension;
 import java.awt.Font;
 import java.awt.GridBagConstraints;
 import java.awt.GridBagLayout;
 import java.awt.Insets;
-import java.awt.event.ActionEvent;
 import java.awt.event.KeyEvent;
 
-import javax.swing.AbstractAction;
 import javax.swing.JButton;
-import javax.swing.JFrame;
 import javax.swing.JLabel;
-import javax.swing.JPanel;
-import javax.swing.JTextField;
 
-public class ArlosGitNotationPrompt {
-  private String message = "";
-  private boolean done = false;
-  private final JPanel panel = new JPanel();
+public class ArlosGitNotationPrompt extends AbstractCommitPrompt {
 
-  public ArlosGitNotationPrompt() {
-    setLayout();
-  }
-
-  private void setLayout() {
+  @Override
+  protected void setLayout() {
     panel.setLayout(new GridBagLayout());
     addCustomCommitPart();
     addQuickKeys();
@@ -63,54 +50,6 @@ public class ArlosGitNotationPrompt {
     createQuickbutton("Delete Clutter", "r   Delete Clutter", KeyEvent.VK_D, position++);
   }
 
-  private void addCustomCommitPart() {
-    {
-      // Commit Message:
-      GridBagConstraints c = new GridBagConstraints();
-      JLabel commitLabel = new JLabel("Commit Message:");
-      c.anchor = GridBagConstraints.FIRST_LINE_START;
-      c.insets = new Insets(10, 10, 0, 0);
-      c.gridx = 1;
-      c.gridwidth = 1;
-      c.gridy = 1;
-      panel.add(commitLabel, c);
-    }
-    JTextField prompt;
-    {
-      // [Commit Message Text]
-      GridBagConstraints c = new GridBagConstraints();
-      prompt = new JTextField("");
-      prompt.setAction(
-          new AbstractAction() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-              doCommit(prompt.getText());
-            }
-          });
-      c.fill = GridBagConstraints.HORIZONTAL;
-      c.anchor = GridBagConstraints.NORTHWEST;
-      c.insets = new Insets(10, 10, 0, 0);
-      c.gridx = 2;
-      c.gridwidth = 3;
-      c.gridy = 1;
-      panel.add(prompt, c);
-    }
-    {
-      // [Commit]
-      var gridBagConstraints = new GridBagConstraints();
-      JButton commit = new JButton("Commit");
-      commit.setMnemonic(KeyEvent.VK_C);
-      commit.setDefaultCapable(true);
-      gridBagConstraints.fill = GridBagConstraints.HORIZONTAL;
-      gridBagConstraints.anchor = GridBagConstraints.NORTHWEST;
-      gridBagConstraints.insets = new Insets(10, 10, 0, 0);
-      gridBagConstraints.gridx = 5;
-      gridBagConstraints.gridwidth = 1;
-      gridBagConstraints.gridy = 1;
-      commit.addActionListener(e -> doCommit(prompt.getText()));
-      panel.add(commit, gridBagConstraints);
-    }
-  }
 
   private void addHelpText(String text, int position) {
     JPanelHelpers.addHelpText(panel, text, position, false);
@@ -146,19 +85,10 @@ public class ArlosGitNotationPrompt {
     panel.add(quick, gridBagConstraints);
   }
 
-  public void doCommit(String text) {
-    this.message = text;
-    this.done = true;
-    Container parent = panel.getParent();
-    while (!(parent instanceof JFrame)) {
-      parent = parent.getParent();
-    }
-    ((JFrame) parent).dispose();
-  }
 
   public static String display() {
     ArlosGitNotationPrompt panel = new ArlosGitNotationPrompt();
-    openInFrame(panel);
+    openInFrame(panel, "Commit...");
     while (!panel.done) {
       try {
         Thread.sleep(500);
@@ -167,25 +97,5 @@ public class ArlosGitNotationPrompt {
       }
     }
     return panel.message;
-  }
-
-  private static void openInFrame(ArlosGitNotationPrompt panel) {
-    JFrame test = new JFrame("Commit...");
-    test.getContentPane().add(panel.panel);
-    test.pack();
-
-    Dimension d = java.awt.Toolkit.getDefaultToolkit().getScreenSize();
-    Dimension w = test.getSize();
-    int dx = (int) w.getWidth();
-    int dy = (int) w.getHeight();
-    int x = (int) ((d.getWidth() - dx) / 2);
-    int y = (int) ((d.getHeight() - dy) / 2);
-    test.setBounds(x, y, dx, dy + 1);
-
-    test.setVisible(true);
-  }
-
-  public JPanel getPanel() {
-    return panel;
   }
 }
